flask-server/requirements.txt
flask
flask_cors
pillow
numpy